audit: complete linear AUDIT.md rollout (QW1-QW8, MH1-MH4, T1-T3, TG0-TG4, CQ1-CQ3, SH1-SH2)
Major split of cli.py (1318 LOC) and studio/widgets/live_modules.py (2451 LOC) into focused sub-packages; extracted OutSim sections from telemetry/protocol/packets.py into packets_outsim.py (re-exported for back-compat). Channel interpretations + Spanish i18n fallbacks moved out of channels.py. Added .github/workflows/test.yml (ruff + pytest on Python 3.13), .pre-commit-config.yaml (pre-commit-hooks + ruff), [tool.mypy] gradual config in pyproject.toml. Narrowed 10 except Exception sites in studio + telemetry modules. Added tests/studio/test_live_modules.py (23 parametrized overlay-window smoke tests) and tests/studio/test_track_map_dock.py (4 dock smoke tests). Repo cleanup: removed scripts/_dep_audit.py, removed loose pytest_*.txt logs, removed unused docs/MANUAL.*.txt duplicates. AUDIT.md fully documents every item with completion notes. Final state: 307 passed / 15 skipped, ruff clean, version locked at 0.4.4.

+++ src/lfs_telemetry/studio/widgets/live_tab.py
@@ -178,7 +178,7 @@ class LiveTab(QWidget):
             )
             try:
                 pct = (
-                    int(round(float(stored) * 100))
+                    round(float(stored) * 100)
                     if stored is not None else 85
                 )
             except (TypeError, ValueError):
@@ -451,8 +451,8 @@ class LiveTab(QWidget):
                 self._vr_status.setText(self._compose_vr_status_text())
             else:
                 err = (
-                    self._vr_mirror._sink.init_error  # noqa: SLF001
-                    if self._vr_mirror._sink is not None  # noqa: SLF001
+                    self._vr_mirror._sink.init_error
+                    if self._vr_mirror._sink is not None
                     else "unavailable"
                 )
                 self._vr_status.setText(
@@ -482,7 +482,7 @@ class LiveTab(QWidget):
         lines: list[str] = [tr("VR mirror active (SteamVR overlay).")]
 
         sink = (
-            self._vr_mirror._sink if self._vr_mirror else None  # noqa: SLF001
+            self._vr_mirror._sink if self._vr_mirror else None
         )
         status = sink.runtime_status() if sink is not None else None
 
@@ -673,7 +673,7 @@ class LiveTab(QWidget):
     # Cleanup
     # ------------------------------------------------------------------
 
-    def closeEvent(self, event) -> None:  # noqa: N802
+    def closeEvent(self, event) -> None:
         if self._vr_mirror is not None:
             self._vr_mirror.shutdown()
             self._vr_mirror = None
